pyproject.toml
# ------------------------
# Project metadata
# ------------------------
[tool.poetry]
name = "notebook_summarizer"
version = "0.1.0"
description = "Summarize Jupyter notebooks into client-friendly reports"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "notebook_summarizer" }]

# ------------------------
# Main runtime dependencies
# ------------------------
[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "*"
jupyter_core = "^5.5.0"
matplotlib = "*"
nbconvert = "^7.10.0"
nbformat = "^5.9.0"
numpy = "*"
pandas = "*"
python-pptx = "^1.0.2"
scikit-learn = "*"
seaborn = "*"
scipy = "^1.15.3"
mistune = "^3.1.3"

# ------------------------
# CLI entry point
# ------------------------
[tool.poetry.scripts]
notebook-summarizer = "notebook_summarizer.cli:main"

# ------------------------
# Development dependencies
# ------------------------
[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = ">=7.0.0"
ruff = ">=0.3.0"

# ------------------------
# Linting configuration
# ------------------------
[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

# ------------------------
# Pytest configuration
# ------------------------
[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]

# ------------------------
# Build backend
# ------------------------
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
